requirements.txt
django-ninja==1.1.0
python-dotenv==1.0.1
